+++ Engine/RumorState.cs
@@ -28,21 +28,26 @@ namespace Exodrifter.Rumor.Engine
 		public List<List<Node>> Consequences { get; private set; }
 
 		/// <summary>
+		/// An event for when dialog is added to the state.
+		/// </summary>
+		public event Action<object, string> OnAddDialog;
+
+		/// <summary>
+		/// An event for when dialog is set in the state.
+		/// </summary>
+		public event Action<object, string> OnSetDialog;
+
+		/// <summary>
+		/// An event for when the state is cleared.
+		/// </summary>
+		public event Action OnClear;
+
+		/// <summary>
 		/// Creates a new Rumor state.
 		/// </summary>
 		public RumorState()
 		{
-			Reset();
-		}
-
-		/// <summary>
-		/// Resets the state.
-		/// </summary>
-		public void Reset()
-		{
-			Dialog = new LazyDictionary<object, string>();
-			Choices = new List<string>();
-			Consequences = new List<List<Node>>();
+			Clear();
 		}
 
 		/// <summary>
@@ -54,6 +59,10 @@ namespace Exodrifter.Rumor.Engine
 		{
 			Dialog.Clear();
 			Dialog[speaker] = dialog;
+
+			if (OnSetDialog != null) {
+				OnSetDialog(speaker, dialog);
+			}
 		}
 
 		/// <summary>
@@ -64,6 +73,10 @@ namespace Exodrifter.Rumor.Engine
 		public void AddDialog(object speaker, string dialog)
 		{
 			Dialog[speaker] += dialog;
+
+			if (OnAddDialog != null) {
+				OnAddDialog(speaker, dialog);
+			}
 		}
 
 		/// <summary>
@@ -96,12 +109,26 @@ namespace Exodrifter.Rumor.Engine
 		}
 
 		/// <summary>
-		/// Removes all current choices.
+		/// Clears the state.
+		/// </summary>
+		public void Clear()
+		{
+			Dialog = new LazyDictionary<object, string>();
+			Choices = new List<string>();
+			Consequences = new List<List<Node>>();
+
+			if (OnClear != null) {
+				OnClear();
+			}
+		}
+
+		/// <summary>
+		/// Clears the choices.
 		/// </summary>
 		public void ClearChoices()
 		{
-			Choices.Clear();
-			Consequences.Clear();
+			Choices = new List<string>();
+			Consequences = new List<List<Node>>();
 		}
 
 		#region Serialization