+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+- Added an `OnDialogAdd` and `OnDialogSet` events to RumorState
+- Added the Clear command
+- Add an `OnWaitForAdvance` and `OnWaitForChoose` event to Rumor
+- Add a `Cancelled` property to Rumor
+
+### Changed
+- AutoAdvance is now a time instead of a toggle, which indicates the amount of
+  time before the Rumor will attempt to automatically advance.
+
 
 ## [0.2.0] - 2016-11-05